requirements.txt
numpy==2.2.2
opencv_python==4.11.0.86
Pillow==11.1.0
streamlit==1.37.1
streamlit==1.41.1
tensorflow==2.18.0
